Release 0.5.8 code (#34)
* Update README.md

replaced a badged
added a badged

* Daily Commit

Previously, serial devices were using RGB LEDs. This was inherited from
Adalight, which uses RGB exclusively. However, this could lead to issues
in the future if a serial single-color device is added. The commit
addresses this concern.

* Daily commit

* changing some characters to line up with the UI.

* Daily commit

* added the ID and ID Grouo separators for linked maps in input plugin
actions

* Daily commit

Fixed and issue in emitter that makes the program to crash if the
contros.ini information is invalid, now an error is displayed and the
game record is ignored.

* 26: Add brightness to elements

Task-Url: https://github.com/meduzapat/LEDSpicer/issues/26

* Added brightness logic

* 27: Update Dates

Task-Url: https://github.com/meduzapat/LEDSpicer/issues/27

Replaced old dates

* 27: Update Dates

Task-Url: https://github.com/meduzapat/LEDSpicer/issues/27

Replaced old dates

* 28: Add more details to the dump

Task-Url: https://github.com/meduzapat/LEDSpicer/issues/28

Improved dump output
added new values to the dump
reduced dump size and duplicated data

* Nightly commit

Doc love

* 32: Research and fix Adaligh protocol issues. (#33)

Task-Url: https://github.com/meduzapat/LEDSpicer/issues/32

Fixed an issue in the header were